+++ admin/includes/fetch_comments.php
@@ -49,7 +49,7 @@
                 <a href='comments.php?source=update_comment&c_id={$comment_id}' class='btn btn-warning'></a>
               </td>";
         echo  "<td class='text-center'>
-                <a href='comments.php?delete={$comment_id}' class='btn btn-danger'></a>
+                <a onClick=\"javascript: return confirm('Are you sure you want to delete?') \" href='comments.php?delete={$comment_id}' class='btn btn-danger'></a>
               </td>";
       echo "</tr>";
     }
